ai-agent/requirements.txt
httpx
python-dotenv
flask
python-json-logger